Move src/packed to src-packed
This was because we were coping all `.js` files in `src/` into `release.zip` as-is.

Simplest fix is to move `src/packed` to `src-packed`.
