+++ src-packed/textAnalytics.js
